requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
numpy==1.20.3
pandas==1.2.4
polygon-api-client==0.1.11
PyJWT==1.7.1
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
tornado==6.1
twilio==6.59.0
tzlocal==2.1
urllib3==1.26.5
websocket-client==1.0.1
websockets==9.1
